services/worker/pyproject.toml
[project]
name = "jyntrix-worker"
version = "1.0.0"
description = "AI Memory Architecture Background Worker"
requires-python = ">=3.11"
dependencies = [
    "arq>=0.26.0",
    "redis>=5.2.0",
    "supabase>=2.11.0",
    "qdrant-client>=1.12.0",
    "sentence-transformers>=3.3.0",
    "google-generativeai>=0.8.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
